#177
-oluşturulan methodlar parçalara ayrıldı.
-Kullanıcı hareketleri bazlı ve kriter bazlı öneri sistemi puanları toplanarak bir tabloya eklendi.

+++ NitelikliBilisim.Core/ViewModels/entity_mirrors/education/EducationVm.cs
@@ -8,9 +8,9 @@ namespace NitelikliBilisim.Core.ViewModels
     public class SuggestedEducationVm : EducationVm
     {
         /// <summary>
-        /// Uygun kriter adedi
+        /// Kriterlere uygunluk puanı
         /// </summary>
-        public int AppropriateCriterionCount { get; set; }
+        public double AppropriateCriterionCount { get; set; }
     }
     public class EducationVm
     {
